fullscreen/api/element-request-fullscreen-two-iframes.html

<!DOCTYPE html>
<title>Element#requestFullscreen() on two elements in different iframes</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<iframe id="a" allowfullscreen></iframe>
<iframe id="b" allowfullscreen></iframe>
<script>
async_test(t => {
  // Request fullscreen on the body elements of both iframes, but in reverse
  // tree order.
  const iframes = [document.getElementById('b'), document.getElementById('a')];

  // Expect two fullscreenchange events in both iframes' documents, in the same
  // order as the requests. Events should also fire on the outer document, but
  // this is not covered by this test.
  const order = [];
  document.onfullscreenchange = t.step_func(() => {
    assert_in_array(document.fullscreenElement, iframes);
    order.push(document.fullscreenElement.id);
    if (order.length == 2) {
      assert_array_equals(order, ['b', 'a'],
                          'fullscreenElement IDs in fullscreenchange events');
      t.done();
    }
  });
  document.onfullscreenerror = t.unreached_func('fullscreenerror event');

  trusted_click(t.step_func(() => {
    iframes[1].contentDocument.body.requestFullscreen();
    iframes[0].contentDocument.body.requestFullscreen();
  }), document.body);
});
</script>
